restructure files, changes style for index, added footer, added contributor
