+++ admin/footer.php
@@ -1,6 +1,7 @@
 <?php
+// footer.php - Global footer include
 ?>
-<link rel="stylesheet" href="styles/footer.css">
+<link rel="stylesheet" href="../styles/footer.css">
 <footer class="site-footer">
   <div class="container">
     <p>&copy; 2025 Group Y, Multimedia University. All rights reserved.</p>
