+++ admin/header.php
@@ -12,20 +12,18 @@ if (session_status() === PHP_SESSION_NONE) {
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>MMU Talent Showcase Portal</title>
-  <link rel="stylesheet" href="styles/header.css">
+  <link rel="stylesheet" href="../styles/header.css">
 </head>
 <body>
 <header>
   <nav class="main-nav">
     <ul>
       <li><a href="index.php">Home</a></li>
-      <li><a href="catalogue.php">Talent Catalogue</a></li>
-      <li><a href="resources.php">Resource Sharing</a></li>
-      <li><a href="faq.php">FAQ</a></li>
+      <li><a href="contributor.php">Contributor</a></li>
     </ul>
     <ul class="user-menu">
       <?php if (isset($_SESSION['user_id'])): ?>
-        <li><a href="profile.php"><img src="assets/user-icon.png" alt="Profile" class="icon"></a></li>
+        <li><a href="profile.php"><img src="../assets/user-icon.png" alt="Profile" class="icon"></a></li>
         <li><a href="logout.php">Logout</a></li>
       <?php else: ?>
         <li><a href="login.php">Login</a></li>
