--- adminindex.php
@@ -1,19 +1,0 @@
-<?php include('adminheader.php'); ?>
-<link rel="stylesheet" href="styles/admin.css">
-
-<main class="admin-container">
-  <div class="content">
-    <!-- Admin Hero Section -->
-    <section id="search-hero">
-      <h1>MMU Talent Showcase Portal<br/> Admin Dashboard</h1>
-    </section>
-      
-<div class="admin-container">
-    <a href="feedback.php" class="admin-card">Users' Feedback</a>
-    <a href="talents.php" class="admin-card">Talent Portfolios</a>
-    <a href="manageusers.php" class="admin-card">Manage Users</a>
-    <a href="announcement.php" class="admin-card">Announcement Management</a>
-    <a href="faq.php" class="admin-card">FAQ Management</a>
-</div>
-
-<?php include 'footer.php'; ?>
