+++ styles/index.css
@@ -1,150 +1,167 @@
 /* styles/index.css */
 
-/* 1. Basic reset & box-sizing */
-* { box-sizing: border-box; margin: 0; padding: 0; }
-
-/* 2. Header & Navigation */
-.main-nav {
-  background-color: #ffffff;
-  border-bottom: 1px solid #e0e0e0;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.75rem 2rem;
-  position: sticky;
-  top: 0;
-  z-index: 1000;
+/* 1. Reset & Box Sizing */
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
 }
 
-.main-nav ul {
-  list-style: none;
-  display: flex;
-  align-items: center;
-}
-
-.main-nav ul li {
-  margin-right: 1.5rem;
-}
-
-.main-nav ul li:last-child {
-  margin-right: 0;
-}
-
-.main-nav a {
-  color: #333333;
-  text-decoration: none;
-  font-weight: 500;
-  padding: 0.5rem;
-  transition: background-color 0.2s ease-in-out;
-}
-
-.main-nav a:hover {
-  background-color: #f5f5f5;
-  border-radius: 4px;
-}
-
-.user-menu li {
-  margin-right: 1rem;
-}
-
-.user-menu li:last-child {
-  margin-right: 0;
-}
-
-.user-menu .icon {
-  width: 24px;
-  height: 24px;
-  display: inline-block;
-}
-
-/* 3. Two-column layout */
+/* 2. Layout */
 .container {
-  display: grid;
-  grid-template-columns: 2fr 1fr;
-  gap: 2rem;
-  padding: 2rem;
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 20px;
+  display: flex;
+  gap: 30px;
+}
+.content {
+  flex: 1;
+  max-width: 800px;
 }
 
-/* 4. Hero/Search styling */
+/* 3. Hero/Search Section */
 #search-hero {
+  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
+  border-radius: 15px;
+  color: #fff;
+  padding: 40px 20px;
   text-align: center;
-  margin-bottom: 2rem;
+  margin-bottom: 40px;
 }
 #search-hero h1 {
-  font-size: 1.8rem;
-  margin-bottom: 1rem;
+  font-size: 2.5rem;
+  font-weight: 700;
+  margin-bottom: 20px;
 }
 #search-hero form {
-  display: inline-flex;
-  width: 100%;
-  max-width: 400px;
+  display: flex;
+  justify-content: center;
+  gap: 10px;
+  flex-wrap: wrap;
+  max-width: 500px;
+  margin: 0 auto;
 }
 #search-hero input {
   flex: 1;
-  padding: 0.5rem;
-  border: 1px solid #ccc;
-  border-right: none;
+  max-width: 400px;
+  padding: 15px 20px;
+  border: none;
+  border-radius: 50px;
+  font-size: 1rem;
+  outline: none;
+  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
 }
 #search-hero button {
-  padding: 0 1rem;
-  border: 1px solid #ccc;
-  background: #eee;
+  padding: 15px 20px;
+  background: rgba(255, 255, 255, 0.2);
+  border: none;
+  border-radius: 50px;
+  color: #fff;
   cursor: pointer;
+  font-size: 1.2rem;
+  transition: background 0.3s ease, transform 0.3s ease;
+}
+#search-hero button:hover {
+  background: rgba(255, 255, 255, 0.3);
+  transform: scale(1.05);
 }
 
-/* 5. Section headings */
+/* 4. Section Headings */
 section h2 {
-  margin-bottom: 1rem;
-  font-size: 1.4rem;
+  font-size: 1.8rem;
+  font-weight: 600;
+  color: #333;
+  margin-bottom: 20px;
 }
 
-/* 6. Grid of cards/placeholders */
+/* 5. Grid & Placeholder Cards */
 .grid {
   display: grid;
-  grid-auto-flow: column;
-  grid-auto-columns: 1fr;
-  gap: 1rem;
+  grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+  gap: 20px;
 }
 .placeholder {
-  background: #ddd;
-  height: 120px;
-  border-radius: 4px;
+  background: #f1f5f9;
+  height: 150px;
+  border-radius: 10px;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.05);
+  transition: transform 0.3s ease;
 }
-.placeholder.small {
-  height: 40px;
-  width: 100%;
+.placeholder:hover {
+  transform: translateY(-5px);
 }
 
-/* 7. Announcements sidebar */
+/* 6. Announcements Sidebar */
+#announcements {
+  width: 300px;
+  flex-shrink: 0;
+}
 #announcements h2 {
-  margin-bottom: 1rem;
+  font-size: 1.5rem;
+  color: #333;
+  margin-bottom: 10px;
 }
 #announcements ul {
   list-style: none;
+  padding: 0;
 }
-#announcements li + li {
-  margin-top: 0.5rem;
+#announcements li {
+  background: #fff;
+  padding: 15px;
+  margin-bottom: 15px;
+  border-radius: 8px;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.05);
+  transition: transform 0.3s ease;
+}
+#announcements li:hover {
+  transform: translateY(-3px);
+}
+#announcements h4 {
+  font-size: 1.1rem;
+  color: #667eea;
+  margin-bottom: 8px;
+}
+#announcements p {
+  font-size: 0.9rem;
+  color: #555;
+  margin-bottom: 10px;
+}
+#announcements small {
+  font-size: 0.8rem;
+  color: #999;
 }
 
-/* 8. Responsive tweaks */
+/* 7. Responsive Breakpoints */
 @media (max-width: 768px) {
-  .main-nav {
-    flex-direction: column;
-    align-items: flex-start;
-    padding: 1rem;
-  }
-  .main-nav ul {
-    flex-direction: column;
-    width: 100%;
-  }
-  .main-nav ul li {
-    margin: 0.5rem 0;
-  }
-
   .container {
-    grid-template-columns: 1fr;
+    flex-direction: column;
   }
   #announcements {
-    margin-top: 2rem;
+    width: 100%;
+    order: -1;
+    margin-bottom: 30px;
+  }
+  #search-hero {
+    padding: 30px 15px;
+  }
+  #search-hero h1 {
+    font-size: 2rem;
+  }
+}
+
+@media (max-width: 480px) {
+  .container {
+    padding: 10px;
+  }
+  #search-hero h1 {
+    font-size: 1.8rem;
+  }
+  #search-hero input {
+    padding: 12px 15px;
+  }
+  #search-hero button {
+    padding: 12px 15px;
+    font-size: 1rem;
   }
 }
